msprechecker/pyproject.toml
# 构建相关
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages]
find = { exclude = ["test*"] }

[tool.setuptools.package-data]
"msprechecker" = ["**/*.yaml", "**/*.cmate"]
"pics" = ["*.png"]

# 运行相关
[project]
name = "msprechecker"
version = "0.3.4"
description = "MindStudio Prechecker"
authors = [
    {name = "AvadaKedavrua", email = "anonymousdev@163.com"},
]
readme = "README.md"
license = {text = "Mulan PSL v2"}
keywords = ["MindStudio", "precheck"]
classifiers = [
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent"
]
dependencies = [
    "pyyaml",
    "psutil",
    "importlib_metadata; python_version == '3.7'",
    "msguard",
    "ply",
    "colorama"
]
requires-python = ">=3.7"

[project.scripts]
msprechecker = "msprechecker.cli:main"

[project.urls]
source = "https://gitcode.com/Ascend/msit/tree/master/msprechecker"
documentation = "https://gitcode.com/Ascend/msit/tree/master/msprechecker"
tracker = "https://gitcode.com/Ascend/msit/issues"

# 可选安装项
[project.optional-dependencies]
test = ['pytest', 'pytest-mocker'] # ut
